Improved simple fields
Changes to be committed:
	modified:   pumpia/module_handling/fields/fields.py
	modified:   pumpia/module_handling/fields/groups.py
	deleted:    pumpia/module_handling/fields/simple.py
	new file:   pumpia/module_handling/fields/values.py
	modified:   pumpia/module_handling/module_collections.py
	modified:   pumpia/module_handling/modules.py
	modified:   pumpia/utilities/typing.py
