+++ pumpia/module_handling/fields/fields.py
@@ -1,25 +1,30 @@
+"""
+Contains simple fields.
+"""
+
 from abc import ABC, abstractmethod
 from collections.abc import Callable
 from typing import TYPE_CHECKING, overload, Self
 from datetime import date
+import tkinter as tk
+from tkinter import ttk
 
-from pumpia.module_handling.fields.simple import (BaseIO,
-                                                  OptionIO,
-                                                  BoolIO,
-                                                  StringIO,
-                                                  IntIO,
-                                                  PercIO,
-                                                  FloatIO,
-                                                  DateIO)
+from pumpia.module_handling.fields.values import (BaseValue,
+                                                  BoolValue,
+                                                  StringValue,
+                                                  IntValue,
+                                                  FloatValue,
+                                                  DateValue)
+from pumpia.widgets.entry_boxes import IntEntry, FloatEntry, DateEntry, PercEntry, DateVar
 
 if TYPE_CHECKING:
     from pumpia.module_handling.modules import BaseModule
 
 
-class Fields:
-    def __init__(self, obj: BaseModule) -> None:
-        self.fields_dict: dict[str, BaseIO] = {}
-        self.obj: BaseModule = obj
+class _Fields:
+    def __init__(self, module: BaseModule) -> None:
+        self.fields_dict: dict[str, BaseField] = {}
+        self.module: BaseModule = module
 
     def __iter__(self):
         for field in self.fields_dict.values():
@@ -29,31 +34,35 @@ class Fields:
 class _FieldsMeta:
     def __init__(self) -> None:
         self.field_types: dict[str, BaseField] = {}
+        self.name: str = ""
+        self.private_name: str = "_"
 
     @property
     def field_names(self) -> list[str]:
         return list(self.field_types.keys())
 
+    def __set_name__(self, owner: type[BaseModule], name: str):
+        self.name = name
+        self.private_name = "_" + name
+
     @overload
-    def __get__(self, obj: BaseModule, owner=None) -> Fields: ...
+    def __get__(self, obj: BaseModule, owner=None) -> _Fields: ...
     @overload
     def __get__(self, obj: None, owner=None) -> Self: ...
 
-    def __get__(self, obj: BaseModule | None, owner=None) -> Fields | Self:
+    def __get__(self, obj: BaseModule | None, owner=None) -> _Fields | Self:
         if obj is None:
             return self
 
         try:
-            return getattr(obj, "fields")
+            return getattr(obj, self.private_name)
         except AttributeError:
-            fields = Fields(obj)
-            for name, field in self.field_types.items():
-                setattr(fields, name, field)
-            setattr(obj, "fields", fields)
+            fields = _Fields(obj)
+            setattr(obj, self.private_name, fields)
             return fields
 
 
-class BaseField[ValT, IOType:BaseIO](ABC):
+class BaseField[ValT, TkVarT:tk.Variable](ABC):
     """
     Base class for input/output handling in modules.
 
@@ -85,89 +94,223 @@ class BaseField[ValT, IOType:BaseIO](ABC):
 
     def __init__(self,
                  initial_value: ValT | Callable[[], ValT],
+                 parent: tk.Misc | None = None,
                  *,
                  verbose_name: str | None = None,
                  label_style: str | None = None,
                  entry_style: str | None = None,
                  hidden: bool = False,
                  read_only: bool = False):
-        self.initial_value: ValT | Callable[[], ValT] = initial_value
-        self.verbose_name: str | None = verbose_name
-        self.label_style: str | None = label_style
-        self.entry_style: str | None = entry_style
-        self.hidden: bool = hidden
-        self.read_only: bool = read_only
-        self.name: str = ""
+        self._name: str | None = verbose_name
+        self._label_style: str | None = label_style
 
-    @property
-    @abstractmethod
-    def IO_base(self) -> type[IOType]:
-        pass
+        self._parent: tk.Misc | None = parent
+        self._labels: list[ttk.Label] = []
+        self._entries: list[ttk.Entry] = []
+        self._label_var: tk.StringVar | None = None
+        self.hidden: bool = hidden
+        self._initial_value: ValT | Callable[[], ValT] = initial_value
+        self._value: BaseValue[ValT, TkVarT] | None = None
+
+        self.read_only: bool = read_only
+        self._entry_style: str | None = entry_style
+
+        self.name: str = ""
+        self.module: BaseModule | None = None
 
     def __set_name__(self, owner: type[BaseModule], name: str):
         self.name = name
+        if self.verbose_name is None:
+            self.verbose_name = name.replace("_", " ").title()
         owner.fields.field_types[name] = self
 
     @overload
     def __get__(self, obj: BaseModule, owner=None) -> ValT: ...
     @overload
-    def __get__(self, obj: Fields, owner=None) -> IOType: ...
-    @overload
-    def __get__(self, obj: None, owner=None) -> Self: ...
+    def __get__(self, obj: None | _Fields, owner=None) -> Self: ...
 
-    def __get__(self, obj: BaseModule | Fields | None, owner=None) -> ValT | IOType | Self:
+    def __get__(self, obj: BaseModule | _Fields | None, owner=None) -> ValT | Self:
         if obj is None:
             return self
-        elif isinstance(obj, Fields):
+
+        elif isinstance(obj, _Fields):
             try:
-                field: IOType = obj.fields_dict[self.name]  # pyright: ignore[reportAssignmentType]
+                return obj.fields_dict[self.name]  # pyright: ignore[reportReturnType]
 
             except KeyError:
-                field = self.IO_base(initial_value=self.initial_value,
-                                     parent=obj.obj,
-                                     verbose_name=self.verbose_name,
-                                     label_style=self.label_style,
-                                     entry_style=self.entry_style,
-                                     hidden=self.hidden,
-                                     read_only=self.read_only)
+                field = type(self)(initial_value=self.initial_value,
+                                   parent=None,
+                                   verbose_name=self.verbose_name,
+                                   label_style=self._label_style,
+                                   entry_style=self._entry_style,
+                                   hidden=self.hidden,
+                                   read_only=self.read_only)
+                field.name = self.name
+                field.module = obj.module
                 obj.fields_dict[self.name] = field
-
-            return field
+                return field
 
         else:
             try:
-                field: IOType = obj.fields.fields_dict[self.name]  # pyright: ignore[reportAssignmentType]
+                return obj.fields.fields_dict[self.name].value
 
             except KeyError:
-                field = self.IO_base(initial_value=self.initial_value,
-                                     parent=obj,
-                                     verbose_name=self.verbose_name,
-                                     label_style=self.label_style,
-                                     entry_style=self.entry_style,
-                                     hidden=self.hidden,
-                                     read_only=self.read_only)
+                field = type(self)(initial_value=self.initial_value,
+                                   parent=None,
+                                   verbose_name=self.verbose_name,
+                                   label_style=self._label_style,
+                                   entry_style=self._entry_style,
+                                   hidden=self.hidden,
+                                   read_only=self.read_only)
+                field.name = self.name
+                field.module = obj
                 obj.fields.fields_dict[self.name] = field
-
-            return field.value
+                return field.value
 
     def __set__(self, obj: BaseModule, value: ValT):
         try:
-            field: IOType = obj.fields.fields_dict[self.name]  # pyright: ignore[reportAssignmentType]
+            field = obj.fields.fields_dict[self.name]
 
         except KeyError:
-            field = self.IO_base(initial_value=self.initial_value,
-                                 parent=obj,
-                                 verbose_name=self.verbose_name,
-                                 label_style=self.label_style,
-                                 entry_style=self.entry_style,
-                                 hidden=self.hidden,
-                                 read_only=self.read_only)
+            field = type(self)(initial_value=self.initial_value,
+                               parent=None,
+                               verbose_name=self.verbose_name,
+                               label_style=self._label_style,
+                               entry_style=self._entry_style,
+                               hidden=self.hidden,
+                               read_only=self.read_only)
+            field.name = self.name
+            field.module = obj
             obj.fields.fields_dict[self.name] = field
 
         field.value = value
 
+    @property
+    def widget(self) -> type[ttk.Entry]:
+        return ttk.Entry
 
-class OptionField[DictValT](BaseField[str, OptionIO[DictValT]]):
+    @property
+    @abstractmethod
+    def value_type(self) -> type[BaseValue[ValT, TkVarT]]:
+        pass
+
+    @property
+    def verbose_name(self) -> str | None:
+        """
+        The verbose name of the input/output.
+        """
+        return self._name
+
+    @verbose_name.setter
+    def verbose_name(self, val: str):
+        self._name = val
+        if self._label_var is not None:
+            self._label_var.set(val)
+
+    @property
+    def value_store(self) -> BaseValue[ValT, TkVarT]:
+        if self._value is None:
+            if self._parent is None:
+                raise ValueError("Parent has not been set")
+            self._value = self.value_type(self._parent, self.initial_value)
+        return self._value
+
+    @property
+    def value_var(self) -> TkVarT:
+        return self.value_store.var
+
+    @property
+    def value(self) -> ValT:
+        """
+        The value of the input/output.
+        """
+        return self.value_store.value
+
+    @value.setter
+    def value(self, val: ValT | BaseValue[ValT, TkVarT]):
+        if isinstance(val, BaseValue):
+            self._value = val
+            self._value_var_setter()
+        else:
+            self.value_store.value = val
+
+    @property
+    def initial_value(self) -> ValT:
+        """
+        The initial value of the IO.
+        """
+        if callable(self._initial_value):
+            return self._initial_value()  # type: ignore
+        else:
+            return self._initial_value
+
+    def new_label(self, parent: tk.Misc) -> ttk.Label:
+        """
+        The label with the name of the input/output.
+        """
+        var = self.label_var
+        if self._label_style is None:
+            label = ttk.Label(parent, textvariable=var)
+        else:
+            label = ttk.Label(parent, style=self._label_style, textvariable=var)
+        self._labels.append(label)
+        return label
+
+    @property
+    def label_var(self) -> tk.StringVar:
+        """
+        The tkinter variable related to `label`.
+        """
+        if self._parent is None:
+            raise ValueError("Parent has not been set")
+        if self.verbose_name is None:
+            raise ValueError("verbose_name has not been set")
+        if self._label_var is None:
+            self._label_var = tk.StringVar(self._parent, value=self.verbose_name)
+        return self._label_var
+
+    def set_parent(self, parent: tk.Misc):
+        """
+        Sets the parent of the input/output.
+
+        Parameters
+        ----------
+        parent : tk.Misc
+            The parent of the input/output.
+        """
+        if self._parent is None:
+            self._parent = parent
+        else:
+            raise ValueError("Parent already set")
+
+    def reset_value(self):
+        """
+        Resets the IO to the initial value.
+        """
+        self.value = self.initial_value
+
+    def new_entry(self, parent: tk.Misc) -> ttk.Entry:
+        """
+        The entry widget of the input.
+        """
+        var = self.value_store.var
+        if self.read_only:
+            state = "readonly"
+        else:
+            state = "normal"
+        if self._entry_style is None:
+            entry = self.widget(parent, textvariable=var, state=state)
+        else:
+            entry = self.widget(parent, textvariable=var, state=state, style=self._entry_style)
+        self._entries.append(entry)
+        return entry
+
+    def _value_var_setter(self):
+        for entry in self._entries:
+            entry.configure(textvariable=self.value_var)
+
+
+class OptionField[DictValT](BaseField[str, tk.StringVar]):
     """
     Represents an option input.
     Has the same attributes and methods as BaseInput unless stated below.
@@ -185,6 +328,7 @@ class OptionField[DictValT](BaseField[str, OptionIO[DictValT]]):
     def __init__(self,
                  options_map: dict[str, DictValT],
                  initial: str | Callable[[], str],
+                 parent: tk.Misc | None = None,
                  *,
                  verbose_name: str | None = None,
                  label_style=None,
@@ -192,142 +336,346 @@ class OptionField[DictValT](BaseField[str, OptionIO[DictValT]]):
                  allow_inv_mapping: bool = False,
                  hidden: bool = False,
                  read_only: bool = False):
-        super().__init__(initial,
+        super().__init__(initial_value=initial,
+                         parent=parent,
                          verbose_name=verbose_name,
                          label_style=label_style,
                          entry_style=entry_style,
                          hidden=hidden,
                          read_only=read_only)
-        self.options_map = options_map
-        self.allow_inv_mapping = allow_inv_mapping
 
-    @property
-    def IO_base(self) -> type[OptionIO[DictValT]]:
-        return OptionIO
+        self._entries: list[ttk.Combobox] = []
+
+        self.allow_inv_mapping: bool = allow_inv_mapping
+        self.options_map: dict[str, DictValT] = options_map
+
+        if self.allow_inv_mapping:
+            self._inv_map: dict[DictValT, str] = {v: k for k, v in self.options_map.items()}
+
+        self.options = list(self.options_map.keys())
+        if initial not in self.options:
+            raise ValueError("initial not in options")
 
     @overload
     def __get__(self, obj: BaseModule, owner=None) -> DictValT: ...
     @overload
-    def __get__(self, obj: Fields, owner=None) -> OptionIO: ...
-    @overload
-    def __get__(self, obj: None, owner=None) -> Self: ...
+    def __get__(self, obj: None | _Fields, owner=None) -> Self: ...
 
-    def __get__(self, obj: BaseModule | Fields | None, owner=None) -> DictValT | OptionIO | Self:
+    def __get__(self, obj: BaseModule | _Fields | None, owner=None) -> DictValT | Self:
         if obj is None:
             return self
 
-        elif isinstance(obj, Fields):
+        elif isinstance(obj, _Fields):
             try:
-                field: OptionIO[DictValT] = obj.fields_dict[self.name]  # pyright: ignore[reportAssignmentType]
+                field: Self = obj.fields_dict[self.name]  # pyright: ignore[reportAssignmentType]
 
             except KeyError:
-
-                field = self.IO_base(initial=self.initial_value,
-                                     options_map=self.options_map,
-                                     parent=obj.obj,
-                                     verbose_name=self.verbose_name,
-                                     label_style=self.label_style,
-                                     entry_style=self.entry_style,
-                                     allow_inv_mapping=self.allow_inv_mapping,
-                                     hidden=self.hidden,
-                                     read_only=self.read_only)
+                field = type(self)(initial=self.initial_value,
+                                   options_map=self.options_map,
+                                   parent=None,
+                                   verbose_name=self.verbose_name,
+                                   label_style=self._label_style,
+                                   entry_style=self._entry_style,
+                                   allow_inv_mapping=self.allow_inv_mapping,
+                                   hidden=self.hidden,
+                                   read_only=self.read_only)
                 obj.fields_dict[self.name] = field
 
             return field
         else:
             try:
-                field: OptionIO[DictValT] = obj.fields.fields_dict[self.name]  # pyright: ignore[reportAssignmentType]
+                field = obj.fields.fields_dict[self.name]  # pyright: ignore[reportAssignmentType]
 
             except KeyError:
-                field = self.IO_base(initial=self.initial_value,
-                                     options_map=self.options_map,
-                                     parent=obj,
-                                     verbose_name=self.verbose_name,
-                                     label_style=self.label_style,
-                                     entry_style=self.entry_style,
-                                     allow_inv_mapping=self.allow_inv_mapping,
-                                     hidden=self.hidden,
-                                     read_only=self.read_only)
+                field = type(self)(initial=self.initial_value,
+                                   options_map=self.options_map,
+                                   parent=None,
+                                   verbose_name=self.verbose_name,
+                                   label_style=self._label_style,
+                                   entry_style=self._entry_style,
+                                   allow_inv_mapping=self.allow_inv_mapping,
+                                   hidden=self.hidden,
+                                   read_only=self.read_only)
                 obj.fields.fields_dict[self.name] = field
 
             return field.value
 
     def __set__(self, obj: BaseModule, value: DictValT | str):
         try:
-            field: OptionIO[DictValT] = obj.fields.fields_dict[self.name]  # pyright: ignore[reportAssignmentType]
+            field: Self = obj.fields.fields_dict[self.name]  # pyright: ignore[reportAssignmentType]
 
         except KeyError:
-            field = self.IO_base(initial=self.initial_value,
-                                 options_map=self.options_map,
-                                 parent=obj,
-                                 verbose_name=self.verbose_name,
-                                 label_style=self.label_style,
-                                 entry_style=self.entry_style,
-                                 allow_inv_mapping=self.allow_inv_mapping,
-                                 hidden=self.hidden,
-                                 read_only=self.read_only)
+            field = type(self)(initial=self.initial_value,
+                               options_map=self.options_map,
+                               parent=None,
+                               verbose_name=self.verbose_name,
+                               label_style=self._label_style,
+                               entry_style=self._entry_style,
+                               allow_inv_mapping=self.allow_inv_mapping,
+                               hidden=self.hidden,
+                               read_only=self.read_only)
             obj.fields.fields_dict[self.name] = field
 
         field.value = value
 
+    @property
+    def widget(self) -> type[ttk.Combobox]:
+        return ttk.Combobox
 
-class BoolField(BaseField[bool, BoolIO]):
+    @property
+    def value_type(self) -> type[StringValue]:
+        return StringValue
+
+    @property
+    def value(self) -> DictValT:
+        """
+        The object mapped to the option as given by `options_map`.
+        Can be set by using the option string or, if `allow_inv_mapping` is True, the object.
+        """
+        value = super().value
+        return self.options_map[value]
+
+    @value.setter
+    def value(self, val: DictValT | str):
+        if val in self.options:
+            super().value = val  # pyright: ignore[reportAttributeAccessIssue]
+        elif self.allow_inv_mapping and val in self._inv_map:
+            val = self._inv_map[val]  # pyright: ignore[reportArgumentType]
+            super().value = val
+        else:
+            raise ValueError("Value not in options")
+
+    @property
+    def entry(self) -> ttk.Combobox:
+        """
+        The Combobox widget of the input.
+        """
+        var = self.value_var
+        if self.read_only:
+            state = "disabled"
+        else:
+            state = "readonly"
+        if self._entry_style is None:
+            entry = self.widget(self._parent,
+                                textvariable=var,
+                                values=self.options,
+                                state=state)
+        else:
+            entry = self.widget(self._parent,
+                                textvariable=var,
+                                values=self.options,
+                                style=self._entry_style,
+                                state=state)
+        self._entries.append(entry)
+        return entry
+
+
+class BoolField(BaseField[bool, tk.BooleanVar]):
     """
     Represents a boolean input.
     Has the same attributes and methods as BaseInput unless stated below.
     """
+
+    def __init__(self,
+                 initial_value: bool | Callable[[], bool] = True,
+                 parent: tk.Misc | None = None,
+                 *,
+                 verbose_name: str | None = None,
+                 label_style: str | None = None,
+                 entry_style: str | None = None,
+                 hidden: bool = False,
+                 read_only: bool = False):
+        super().__init__(initial_value,
+                         parent=parent,
+                         verbose_name=verbose_name,
+                         label_style=label_style,
+                         entry_style=entry_style,
+                         hidden=hidden,
+                         read_only=read_only)
+        self._entries: list[ttk.Checkbutton] = []
+
     @property
-    def IO_base(self) -> type[BoolIO]:
-        return BoolIO
+    def widget(self) -> type[ttk.Checkbutton]:
+        return ttk.Checkbutton
+
+    @property
+    def value_type(self) -> type[BoolValue]:
+        return BoolValue
+
+    def _value_var_setter(self):
+        for entry in self._entries:
+            entry.configure(variable=self.value_var)
+
+    def new_entry(self, parent: tk.Misc) -> ttk.Checkbutton:
+        """
+        The Checkbutton widget of the input.
+        """
+        var = self.value_var
+        if self.read_only:
+            state = "readonly"
+        else:
+            state = "normal"
+        if self._entry_style is None:
+            entry = self.widget(self._parent, variable=var, state=state)
+        else:
+            entry = self.widget(self._parent, variable=var, state=state, style=self._entry_style)
+        self._entries.append(entry)
+        return entry
 
 
-class StringField(BaseField[str, StringIO]):
+class StringField(BaseField[str, tk.StringVar]):
     """
     Represents a string input.
     Has the same attributes and methods as BaseInput unless stated below.
     """
+
+    def __init__(self,
+                 initial_value: str | Callable[[], str] = "",
+                 parent: tk.Misc | None = None,
+                 *,
+                 verbose_name: str | None = None,
+                 label_style: str | None = None,
+                 entry_style: str | None = None,
+                 hidden: bool = False,
+                 read_only: bool = False):
+        super().__init__(initial_value,
+                         parent=parent,
+                         verbose_name=verbose_name,
+                         label_style=label_style,
+                         entry_style=entry_style,
+                         hidden=hidden,
+                         read_only=read_only)
+
     @property
-    def IO_base(self) -> type[StringIO]:
-        return StringIO
+    def value_type(self) -> type[StringValue]:
+        return StringValue
 
 
-class IntField(BaseField[int, IntIO]):
+class IntField(BaseField[int, tk.IntVar]):
     """
     Represents an integer input.
     Has the same attributes and methods as BaseInput unless stated below.
     """
+
+    def __init__(self,
+                 initial_value: int | Callable[[], int] = 0,
+                 parent: tk.Misc | None = None,
+                 *,
+                 verbose_name: str | None = None,
+                 label_style: str | None = None,
+                 entry_style: str | None = None,
+                 hidden: bool = False,
+                 read_only: bool = False):
+        super().__init__(initial_value,
+                         parent=parent,
+                         verbose_name=verbose_name,
+                         label_style=label_style,
+                         entry_style=entry_style,
+                         hidden=hidden,
+                         read_only=read_only)
+
     @property
-    def IO_base(self) -> type[IntIO]:
-        return IntIO
+    def widget(self) -> type[IntEntry]:
+        return IntEntry
+
+    @property
+    def value_type(self) -> type[IntValue]:
+        return IntValue
 
 
-class PercField(BaseField[float, PercIO]):
+class PercField(BaseField[float, tk.IntVar]):
     """
     Represents a percentage input.
     Has the same attributes and methods as BaseInput unless stated below.
     """
+
+    def __init__(self,
+                 initial_value: float | Callable[[], float] = 0,
+                 parent: tk.Misc | None = None,
+                 *,
+                 verbose_name: str | None = None,
+                 label_style: str | None = None,
+                 entry_style: str | None = None,
+                 hidden: bool = False,
+                 read_only: bool = False):
+        if isinstance(initial_value, float) and initial_value > 100:
+            raise ValueError("Initial value must be less that 100")
+        super().__init__(initial_value,
+                         parent=parent,
+                         verbose_name=verbose_name,
+                         label_style=label_style,
+                         entry_style=entry_style,
+                         hidden=hidden,
+                         read_only=read_only)
+
     @property
-    def IO_base(self) -> type[PercIO]:
-        return PercIO
+    def widget(self) -> type[PercEntry]:
+        return PercEntry
+
+    @property
+    def value_type(self) -> type[FloatValue]:
+        return FloatValue
 
 
-class FloatField(BaseField[float, FloatIO]):
+class FloatField(BaseField[float, tk.DoubleVar]):
     """
     Represents a float input.
     Has the same attributes and methods as BaseInput unless stated below.
     """
 
+    def __init__(self,
+                 initial_value: float | Callable[[], float] = 0,
+                 parent: tk.Misc | None = None,
+                 *,
+                 verbose_name: str | None = None,
+                 label_style: str | None = None,
+                 entry_style: str | None = None,
+                 hidden: bool = False,
+                 read_only: bool = False):
+        super().__init__(initial_value,
+                         parent=parent,
+                         verbose_name=verbose_name,
+                         label_style=label_style,
+                         entry_style=entry_style,
+                         hidden=hidden,
+                         read_only=read_only)
+
     @property
-    def IO_base(self) -> type[FloatIO]:
-        return FloatIO
+    def widget(self) -> type[FloatEntry]:
+        return FloatEntry
+
+    @property
+    def value_type(self) -> type[FloatValue]:
+        return FloatValue
 
 
-class DateField(BaseField[date, DateIO]):
+class DateField(BaseField[date, DateVar]):
     """
     Represents a date input.
     Has the same attributes and methods as BaseInput unless stated below.
     """
 
+    def __init__(self,
+                 initial_value=date.today(),
+                 parent: tk.Misc | None = None,
+                 *,
+                 verbose_name: str | None = None,
+                 label_style: str | None = None,
+                 entry_style: str | None = None,
+                 hidden: bool = False,
+                 read_only: bool = False):
+        super().__init__(initial_value,
+                         parent=parent,
+                         verbose_name=verbose_name,
+                         label_style=label_style,
+                         entry_style=entry_style,
+                         hidden=hidden,
+                         read_only=read_only)
+
     @property
-    def IO_base(self) -> type[DateIO]:
-        return DateIO
+    def widget(self) -> type[DateEntry]:
+        return DateEntry
+
+    @property
+    def value_type(self) -> type[DateValue]:
+        return DateValue
