+++ pumpia/module_handling/fields/groups.py
@@ -1,10 +1,53 @@
 """
 Contains groupings of inputs/outputs.
 """
-from pumpia.module_handling.fields.simple import BaseIO
+from typing import overload, Self
+from pumpia.module_handling.fields.fields import BaseField
+from pumpia.module_handling.module_collections import BaseCollection
 
 
-class IOGroup:
+class _FieldGroups:
+    def __init__(self, obj: BaseCollection) -> None:
+        self.groups_dict: dict[str, FieldGroup] = {}
+        self.obj: BaseCollection = obj
+
+    def __iter__(self):
+        for module in self.groups_dict.values():
+            yield module
+
+
+class _FieldGroupsMeta:
+    def __init__(self) -> None:
+        self.groups: dict[str, FieldGroup] = {}
+        self.name: str = ""
+        self.private_name: str = "_"
+
+    @property
+    def group_names(self) -> list[str]:
+        return list(self.groups.keys())
+
+    def __set_name__(self, owner: type[BaseCollection], name: str):
+        self.name = name
+        self.private_name = "_" + name
+
+    @overload
+    def __get__(self, obj: BaseCollection, owner=None) -> _FieldGroups: ...
+    @overload
+    def __get__(self, obj: None, owner=None) -> Self: ...
+
+    def __get__(self, obj: BaseCollection | None, owner=None) -> _FieldGroups | Self:
+        if obj is None:
+            return self
+
+        try:
+            return getattr(obj, self.private_name)
+        except AttributeError:
+            groups = _FieldGroups(obj)
+            setattr(obj, self.private_name, groups)
+            return groups
+
+
+class FieldGroup:
     """
     Represents a group of linked input / output objects.
     IOs should only be a member of one group.
@@ -19,13 +62,12 @@ class IOGroup:
     linked_ios: list[BaseIO]
     """
 
-    def __init__(self, linked_ios: list[BaseIO]):
-        var_type = linked_ios[0].var_type
-        for vt in linked_ios[1:]:
-            if vt.var_type is not var_type:
-                raise ValueError("IOs not the same variable type")
+    def __init__(self, fields: list[BaseField]):
+        self.name: str = ""
+        self.module_fields: list[tuple[str, str]] = [(field.module.name, field.name)
+                                                     for field in fields
+                                                     if field.module is not None]
 
-        self.linked_ios: list[BaseIO] = linked_ios
-
-        for io in self.linked_ios[1:]:
-            io.value_var = self.linked_ios[0].value_var
+    def __set_name__(self, owner: type[BaseCollection], name: str):
+        self.name = name
+        owner.field_groups.groups[name] = self
