+++ pumpia/utilities/typing.py
@@ -3,4 +3,4 @@
 from typing import Literal
 
 type SideType = Literal["top", "bottom", "left", "right"]
-type DirectionType = Literal["Horizontal", "h", "H", "Vertical", "v", "V"]
+type DirectionType = Literal["Horizontal", "horizontal", "h", "H", "Vertical", "vertical", "v", "V"]
